+++ TestWebsite/Client/Scripts/Pages/Test/TabManager.tsx
@@ -11,6 +11,7 @@ import {
 
 interface TabContentProps {
     onReady(): void
+    loadImmediately?: boolean
 }
 
 interface TabContentState {
@@ -18,17 +19,24 @@ interface TabContentState {
 }
 
 class TabContent extends React.Component<TabContentProps, TabContentState> {
+    static defaultProps: Pick<TabContentProps, 'loadImmediately'> = {
+        loadImmediately: false
+    }
+
     state: TabContentState = { dataLoaded: false }
     timer?: number
 
     componentDidMount() {
-        this.timer = setTimeout(() => {
-            this.setState({
-                dataLoaded: true
-            })
+        this.timer = setTimeout(
+            () => {
+                this.setState({
+                    dataLoaded: true
+                })
 
-            this.props.onReady()
-        }, 2000)
+                this.props.onReady()
+            },
+            !this.props.loadImmediately ? 2000 : 0
+        )
     }
 
     componentWillUnmount() {
@@ -76,11 +84,13 @@ interface Readiness {
 
 interface PageState {
     readiness: Readiness
+    displaySingleTab: boolean
 }
 
 export class Page extends React.Component<PageProps, PageState> {
     state: PageState = {
-        readiness: { a: false, b: false, c: false }
+        readiness: { a: false, b: false, c: false },
+        displaySingleTab: false
     }
 
     get tab() {
@@ -119,35 +129,78 @@ export class Page extends React.Component<PageProps, PageState> {
 
     render() {
         const { ready } = this.props
-        const { readiness } = this.state
+        const { readiness, displaySingleTab } = this.state
 
         return (
             <div className={ready ? '' : 'd-none'}>
-                <TabManager tabs={tabs}>
-                    {[
-                        [
-                            TabName.A,
-                            !readiness.a,
-                            <TabContent onReady={() => this.onChildReady({ a: true })}>
-                                A
-                            </TabContent>
-                        ],
-                        [
-                            TabName.B,
-                            !readiness.b,
-                            <TabContent onReady={() => this.onChildReady({ b: true })}>
-                                B
-                            </TabContent>
-                        ],
-                        [
-                            TabName.C,
-                            !readiness.c,
-                            <TabContent onReady={() => this.onChildReady({ c: true })}>
-                                C
-                            </TabContent>
-                        ]
-                    ]}
-                </TabManager>
+                <div className="mb-5">
+                    <TabManager tabs={tabs}>
+                        {[
+                            [
+                                TabName.A,
+                                !readiness.a,
+                                <TabContent
+                                    onReady={() => this.onChildReady({ a: true })}
+                                >
+                                    A
+                                </TabContent>
+                            ],
+                            [
+                                TabName.B,
+                                !readiness.b,
+                                <TabContent
+                                    onReady={() => this.onChildReady({ b: true })}
+                                >
+                                    B
+                                </TabContent>
+                            ],
+                            [
+                                TabName.C,
+                                !readiness.c,
+                                <TabContent
+                                    onReady={() => this.onChildReady({ c: true })}
+                                >
+                                    C
+                                </TabContent>
+                            ]
+                        ]}
+                    </TabManager>
+                </div>
+                <div>
+                    <h4>TabManager with only one tab</h4>
+                    <div className="form-check form-check-inline mt-2 mb-3">
+                        <input
+                            id="display-single-tab-checkbox"
+                            type="checkbox"
+                            className="form-check-input"
+                            checked={displaySingleTab}
+                            onChange={() =>
+                                this.setState({ displaySingleTab: !displaySingleTab })
+                            }
+                        />
+                        <label
+                            htmlFor="display-single-tab-checkbox"
+                            className="form-check-label"
+                        >
+                            Display single tab
+                        </label>
+                    </div>
+                    <TabManager
+                        tabs={[tabs[0]]}
+                        urlParamName="tab2"
+                        displaySingleTab={displaySingleTab}
+                    >
+                        {[
+                            [
+                                TabName.A,
+                                !readiness.a,
+                                <TabContent onReady={() => {}} loadImmediately>
+                                    A
+                                </TabContent>
+                            ]
+                        ]}
+                    </TabManager>
+                </div>
             </div>
         )
     }